requirements.txt
streamlit
torch
transformers
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pymupdf
pillow
scikit-learn
python-dotenv